InfoSaúde_Caruaru_V1.0_Inventory_StoreRoom

+++ app/Http/Controllers/Inventory/InventoryStoreRoomController.php
@@ -3,9 +3,13 @@
 namespace App\Http\Controllers\Inventory;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Inventory\StoreInventoryStoreRoomPermissionRequest;
 use App\Http\Requests\Inventory\StoreInventoryStoreRoomRequest;
 use App\Http\Requests\Inventory\UpdateInventoryStoreRoomRequest;
 use App\Models\Inventory\InventoryStoreRoom;
+use App\Models\Inventory\InventoryStoreRoomItem;
+use App\Models\Inventory\InventoryStoreRoomPermission;
+use App\Models\Inventory\InventoryStoreRoomPermissionUser;
 use Illuminate\Http\Request;
 
 class InventoryStoreRoomController extends Controller
@@ -37,10 +41,21 @@ class InventoryStoreRoomController extends Controller
      */
     public function destroy(InventoryStoreRoom $inventoryStoreRoom)
     {
-        //        
-        $inventoryStoreRoom->delete();
+        //
+        $dbItem = InventoryStoreRoomItem::where('inventory_store_room_id',$inventoryStoreRoom->id)->count();
 
-        return redirect()->back()->with('success','Almoxarifado excluído com sucesso');
+        if ($dbItem <= 0) {
+            $inventoryStoreRoom->delete();
+    
+            return redirect()->back()->with('success','Almoxarifado excluído com sucesso');
+        } else {
+
+            $inventoryStoreRoom->update([
+                'status' => FALSE
+            ]);
+
+            return redirect()->back()->with('error','Almoxarifado contém itens cadastrados, realizada somente a desativação');
+        }        
     }
 
     /**
@@ -52,5 +67,24 @@ class InventoryStoreRoomController extends Controller
         $inventoryStoreRoom->update($request->all());
 
         return redirect()->back()->with('success','Almoxarifado alterado com sucesso');
+    }    
+
+    /**
+     * Update the specified resource in storage.
+     */
+    public function permission(StoreInventoryStoreRoomPermissionRequest $request)
+    {
+        //
+        $dbPermission = InventoryStoreRoomPermission::where('user_id',$request['id'])
+        ->where('inventory_store_room_id',$request['inventory_store_room_id'])
+        ->count();
+
+        if ($dbPermission <= 0) {
+            InventoryStoreRoomPermission::create($request->all());
+            
+            return redirect()->back()->with('success','Permissão aplicada com sucesso');
+        } else {
+            return redirect()->back()->with('error','Usuário com permissão atribuida ao almoxarifado selecionado');
+        }
     }
 }
